pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dmm"
version = "1.0.0"
description = "Deep Markov Model for probabilistic time series modeling"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "SilviaCleva", email = "email@example.com"}
]
maintainers = [
    {name = "SilviaCleva", email = "email@example.com"}
]
keywords = [
    "deep-learning",
    "probabilistic-programming",
    "time-series",
    "variational-inference",
    "pyro",
    "pytorch",
    "state-space-models",
    "forecasting",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Information Analysis",
    "Typing :: Typed",
]
requires-python = ">=3.10"
dependencies = [
    "torch>=2.0.0",
    "pyro-ppl>=1.9.1",
    "pandas>=2.0.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.13.0",
    "scikit-learn>=1.3.0",
    "numpy>=1.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "mypy>=1.5.0",
    "ruff>=0.1.0",
    "pre-commit>=3.4.0",
    "pandas-stubs>=2.0.0",
]

[project.urls]
Homepage = "https://github.com/SilviaCleva/dmm"
Documentation = "https://github.com/SilviaCleva/dmm#readme"
Repository = "https://github.com/SilviaCleva/dmm.git"
Issues = "https://github.com/SilviaCleva/dmm/issues"
Changelog = "https://github.com/SilviaCleva/dmm/blob/main/CHANGELOG.md"

[tool.setuptools.packages.find]
where = ["."]
include = ["dmm*"]

[tool.ruff]
target-version = "py310"
line-length = 88
exclude = [
    ".git",
    ".mypy_cache",
    ".ruff_cache",
    ".venv",
    "__pycache__",
    "build",
    "dist",
]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
    "B905",   # zip without strict parameter
    "ARG001", # unused function argument
]

[tool.ruff.lint.isort]
known-first-party = ["dmm"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
strict_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
show_error_codes = true

[[tool.mypy.overrides]]
module = [
    "pyro.*",
    "torch.*",
    "matplotlib.*",
    "seaborn.*",
    "sklearn.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]

[tool.coverage.run]
source = ["dmm"]
branch = true
omit = [
    "*/tests/*",
    "*/__init__.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
]
